Add CMOCEAN colormaps
Here we add to the collection the celebrated cmocean colormaps, adapting
the matlab implementation from one of the original authors, @chadagreen!

> original matlab repo: https://github.com/chadagreene/cmocean
> original documentation: https://matplotlib.org/cmocean/

The code is provided under the original MIT license. It's been modified
to comply with our palette/paletteshow syntax. No paletteditor available
at the moment and no plans for a near future implementation.

TODO: add a proper README